chore(Storybook): Algolia support
