+++ src/Welcome.stories.mdx
@@ -12,8 +12,6 @@ import Welcome from './Welcome';
 
 <Welcome id="diveinto" />
 
-<Algolia />
-
 <center>
 	<h1>Welcome</h1>
 	<p>
@@ -24,6 +22,8 @@ import Welcome from './Welcome';
 		at Talend.
 	</p>
 </center>
+
+<Algolia />
 
 <Grid style={{ margin: '5rem 0' }}>
 	<Card
